Add support for redis sentinels for celery broker and backend
Celery supports redis sentinels.
We just need to allow users to set the right options.
The transport_options need a dict.
I didn't want to hardcode the possible keys in the dict.
Therfore we need users to supply the whole dict.
The easiest solution I found was using json loads.
But this is not optimal.